new bullets

+++ src/main/java/com/nindybun/usefulguns/items/guns/AbstractGun.java
@@ -72,7 +72,7 @@ public class AbstractGun extends Item {
             if (ammo.equals(stack, false)) {
                 AbstractBullet abstractBullet = (AbstractBullet) (ammo.getItem() instanceof AbstractBullet ? ammo.getItem() : ModItems.FLINT_BULLET);
                 BulletEntity bulletEntity = abstractBullet.createProjectile(world, ammo, player);
-                bulletEntity.setDamage((bulletEntity.getDamage()*this.damageMultiplier+this.bonusDamage));
+                bulletEntity.setDamage((float) (bulletEntity.getDamage()*this.damageMultiplier+this.bonusDamage));
                 bulletEntity.shootFromRotation(player, player.getRotationVector().x, player.getRotationVector().y, 0, (float) getProjectileSpeed(gun), 0);
                 world.addFreshEntity(bulletEntity);
                 return i;
